Should be fix. if your users that need to be alocated % user per group is not 0. it now will round the needed groups up If you wanting to divide 4 users over groups of 3 you get 2 groups of 2 now rather then 1 group of 4

+++ src/main/java/Models/UserAllocation.java
@@ -102,7 +102,7 @@ public class UserAllocation {
      * @param toAllocate 
      */
     private void putUsersInGroup(ArrayList<User> toAllocate){
-        ArrayList<Group> newGroups = createNewGroups((int) Math.ceil(toAllocate.size() / maxUsers));
+        ArrayList<Group> newGroups = createNewGroups((int) Math.ceil(toAllocate.size() / (double)maxUsers));
         divideUsersOverGroups(newGroups, toAllocate);
     }
